Move old theme in the asset folder
