+++ src/assets/oldTheme/icons.ts
